fix: edited hero and moved some ticket modals
