+++ app/components/shared/ticket-card/accent-decoration.tsx
